добавил шаблон home в первом приближении
