+++ moscowliuda/settings.py
@@ -14,7 +14,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "authorization.apps.AuthorizationConfig",
-    "moscowliuda_main.apps.MoscowliudaMainConfig",
+    "moscowliuda_views.apps.MoscowliudaMainConfig",
     "goods.apps.GoodsConfig",
 ]
 
@@ -79,5 +79,6 @@ STATIC_ROOT = os.path.join(BASE_DIR, 'static')
 MEDIA_ROOT = os.path.join(BASE_DIR, 'media')
 STATIC_URL = '/static/'
 MEDIA_URL = '/media/'
+
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 LOGIN_REDIRECT_URL = 'home'
