+++ moscowliuda/urls.py
@@ -14,14 +14,16 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
+from django.conf import settings
+from django.conf.urls.static import static
 
 from django.contrib import admin
 from django.urls import path, include
 
-from moscowliuda_main.views import home
+from moscowliuda_views.views import HomeView
 
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("auth/", include("django.contrib.auth.urls")),
-    path('', home, name='home'),
-]
+    path('', HomeView.as_view(), name='home'),
+] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
