+++ moscowliuda_views/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class MoscowliudaMainConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'moscowliuda_views'
