+++ templates/project_list.html
@@ -1,0 +1,5 @@
+<ul>
+    {% for project in projects %}
+        <li>{{ project.title }}</li>
+    {% endfor %}
+</ul>
